weather_api/requirements.txt
fastapi
uvicorn
httpx
redis
python-dotenv
slowapi